requirements.txt
Django==1.7.1
MySQL-python==1.2.5
Pillow==2.7.0
Pygments==2.0.1
bpython==0.13.1
django-appconf==0.6
django-bower==5.0.1
django-braces==1.4.0
django-compressor==1.4
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.2
django-grappelli==2.6.3
django-libsass==0.2
easy-thumbnails==2.2
libsass==0.6.2
pytz==2014.10
requests==2.5.1
six==1.9.0
sqlparse==0.1.14
uWSGI==2.0.9
